UrlForm: adjust URLs based on if it starts with /
After a URL is parsed with Uri.parse(), if the scheme is missing, http:// is
prepended unconditionally, and a Format exception is expected. This exception
can never happen, as any valid Uri that is missing a scheme with the scheme
added will be still a valid Uri.

So instead of doing this, we just check the amount of slashes the URL begins
with to construct a more meaningful/correct fixed URL when adding the scheme.