Inject http.Client dependency via a provider
The http.Client is now created outside of the app (in main), and propagated via an
InheritedWidget called HttpClientProvider. The UrlForm now passed the provided
http.Client to the NoClickService instead.

This will help testing making it possible to mock the client at all levels (from
individual deep widgets to the whole app). It also leaves the environmental
hacks to create the client to the main, which is where these decision need to
be actually made.