Remove declaration of intent from the LICENSE file
Having it there is good from one side because anyone caring about the
license will read the declaration of intent, but on the other hand it
might be hairy legally as I don't know if this text could be considered
part of the license or conflict with the restriction to modify the
license text itself.

The declaration of intent is moved to the README (which should probably
be read by anyone caring about the license) and still makes it quite
visible. A nice side effect is GitHub will detect the license correctly.

The LICENSE file was renamed to LICENSE.md so the formatting is also
interpreted correctly.

Fixes #14.