Add script to share CI and devel checks
The new tool/ci now does more or less the same tests as the .github/workflows/ci
(but it doesn't build, as it is less likely to fail, or deploy of course), so
while doing development is easy to just call tool/ci to run a check. The new
script is also called from .githooks/pre-commit, so if git hooks are enabled, it
will be checked for each commit.

The tool take optional commands to run, by default ci is used which performs all
the tests. Other available commands are analyze, format, check, test, test-*,
etc.