codecov: Add patch configuration and relax settings
We allow to reduce coverage by a 0.5% and require only a minimum of 98%
for the whole project, as sometimes it is very impractical to test
everything (and sometimes flutter coverage reporting is faulty and
reports less coverage than it should).

In particular with patches, as the project starts growing, it can happen
that a small patch have some code that is very impractical to test
fully, so we even allow a 90% coverage for patches, as long as the whole
project is not affected very badly.

Signed-off-by: Leandro Lucarella <luca@llucax.com>