Migrate to null-safety
There are a few more, technically unrelated but minimal, changes, but
they are minimal and it would have been hard to make the code work
without them.

There is more work to adapt things to null-safety to use it more
idiomatically.

A big change needed because of this is now mockito mocks use code
generation, which is quite annoying. Generated files are not added to
the repo; to generate them use:

  flutter pub run build_runner build

This is being run by the CI automatically.

Signed-off-by: Leandro Lucarella <luca@llucax.com>